fix(tests): restore worldenergydata CI readiness (#393)
Squash merge CI readiness fixes after all required checks passed.

+++ src/worldenergydata/bsee/analysis/well_api12.py
@@ -532,9 +532,11 @@ class WellAPI12:
             api12_dir_survey_df = directional_surveys[
                 directional_surveys.API12 == api12
             ].copy()
-            api12_dir_survey_df["az"] = 0
-            api12_dir_survey_df["inc"] = 0
-            api12_dir_survey_df["md"] = api12_dir_survey_df["SURVEY_POINT_MD"]
+            api12_dir_survey_df["az"] = 0.0
+            api12_dir_survey_df["inc"] = 0.0
+            api12_dir_survey_df["md"] = api12_dir_survey_df["SURVEY_POINT_MD"].astype(
+                float
+            )
 
             for df_row in range(0, len(api12_dir_survey_df)):
                 WELL_N_S_CODE = api12_dir_survey_df.iloc[df_row]["WELL_N_S_CODE"]
